[iOS] feat: IconView, PaletteView 분리
