+++ iOS/DaDaIkSeon/DaDaIkSeon/Main/View/TokenEditView.swift
@@ -12,7 +12,7 @@ struct TokenEditView: View {
     @State var isEditing = false
     @State var text = ""
     @State private var segmentedMode = 0
-    @State private var palette = ["색상", "아이콘"]
+    @State private var segmentList = ["색상", "아이콘"]
     @Environment(\.presentationMode) var mode: Binding<PresentationMode>
     
     var body: some View {
@@ -45,8 +45,8 @@ struct TokenEditView: View {
             Spacer()
             VStack(spacing: 40) {
                 Picker("palette", selection: $segmentedMode) {
-                    Text(palette[0]).tag(0)
-                    Text(palette[1]).tag(1)
+                    Text(segmentList[0]).tag(0)
+                    Text(segmentList[1]).tag(1)
                 }
                 .pickerStyle(SegmentedPickerStyle())
                 .onTapGesture {
@@ -75,7 +75,7 @@ struct TokenEditView: View {
                 }
             })
             .frame(width: 85)
-            .padding(.vertical, 10.0)
+            .padding(.vertical, 10)
             .background(LinearGradient.mint)
             .cornerRadius(15)
             }
@@ -104,57 +104,6 @@ struct TokenEditView: View {
             hideKeyboard()
         }
     }
-}
-
-struct IconView: View {
-    
-    private var columns = [GridItem(.flexible()),
-                           GridItem(.flexible()),
-                           GridItem(.flexible()),
-                           GridItem(.flexible()),
-                           GridItem(.flexible())]
-    
-    // 아이콘 이미지 저장방법은?
-    
-    var body: some View {
-        LazyVGrid(columns: columns) {
-            ForEach(1...15, id: \.self) { _ in
-                Circle()
-                    .foregroundColor(.mint1)
-                    .frame(width: 35, height: 35, alignment: .center)
-                    .padding(5)
-            }
-        }
-        .padding(.bottom, 16)
-    }
-    
-}
-
-struct PaletteView: View {
-    
-    private var columns = [GridItem(.flexible()),
-                           GridItem(.flexible()),
-                           GridItem(.flexible())]
-    
-    // Color 저장 방법은?
-    
-    private var color: [Color] = [
-        Color.blue1, Color.brown1, Color.pink1,
-        Color.navy1, Color.salmon1, Color.mint1
-    ]
-    
-    var body: some View {
-        LazyVGrid(columns: columns) {
-            ForEach(0...5, id: \.self) { index in
-                Circle()
-                    .foregroundColor(color[index])
-                    .frame(width: 55, height: 55, alignment: .center)
-                    .padding(8)
-            }
-        }
-        .padding(.bottom, 16)
-    }
-    
 }
 
 struct TokenEditView_Previews: PreviewProvider {
